Edit the styles of input pseudo-classes
Requirements were removed from some of the inputs so that their pseudo styles could be modified differently for this project. Only passwords have requirements and the borders are green when valid and red when invalid. In other inputs, the focus makes the borders blue. In addition, hover styles were added to the button.